extract docs processing cluster to doc_processor.py
Pure code move: 8 functions from manifest.py to a new
dbt/parser/doc_processor.py sibling module. No logic changes.

Functions moved:
- _get_doc_blocks
- _process_docs_for_node
- _process_docs_for_source
- _process_docs_for_macro
- _process_docs_for_exposure
- _process_docs_for_metrics
- _process_docs_for_semantic_model
- _process_docs_for_saved_query

manifest.py now imports the 7 public-facing _process_docs_for_*
symbols from doc_processor; _get_doc_blocks is an internal helper
used only within doc_processor.

Imports removed from manifest.py:
- jinja2.nodes.Call / Const (only needed in _get_doc_blocks)
- dbt_common.clients.jinja.parse (only needed in _get_doc_blocks)
- dbt.clients.jinja.get_rendered (only needed in _process_docs_for_*)

DocsContextCallback type alias and the now-unused ResultNode import
remain in manifest.py as dead code; both will be removed in the
follow-up cleanup commit.

Test updated: test_get_doc_blocks.py import path updated to reflect
the new module location.

CodeScene delta skipped on this commit: the new file has the same
high CC as the original manifest.py cluster. Follow-up commit
consolidates duplication and will leave the file in a clean state.
